From chesstab-4.3.tar.gz uploaded to solentware.core.uk on 2020-09-02.
Adjustments to fit the new interface provided by pgn-read-2.0.

Move numbers are displayed by default, but can be switched on or off.

Recursive Annotation Variations (RAV) and comments always start on a new
line when displayed.

PGN errors are put inside a comment which continues to the end of the
RAV, or the end of game if the error is not in a RAV.  If the error is
fixed by editing the game the sequence is restored to the game outside
the comment.

The list of games matching the current position in the active game shows
just the move into, and the move out of, the position along with any '('
and ')' marking RAVs.  Colour highlighting to show difference from the
active game is retained.  This summary game score appears before the
player names rather than after.

On Microsoft Windows when using the pointer to enable scrolling of the
list of games matching a position, it is necessary to double-click rather
than click when done over game details.