From chesstab-4.3.2.tar.gz uploaded to solentware.co.uk on 2021.01.02.
Adjustments to fit new version of dependency pgn-read.

Fix problem wrapping '{..}' comments inside an '{Error}' comment when an
error is found in a game in a PGN file.

Fix problem wrapping '--' movetext in the main line of a game in an
{Error} comment. '--' movetext in Recursive Annotion Variations '(...)' is
not affected.

Fix problems displaying records generated from text with little resemblance
to valid PGN.  These problems usually caused ChessTab to crash.