Rewrite event bindings code and fix some problems.
One problem mentioned because it seemed impossible to fix without rewriting
the event bindings code.  A program crash occurred when a move is done, or
undone, with the 'Down' or 'Up' key after backing out of variation selection
with pointer action on the board or game score.

Event bindings are changed from one state to another by undoing all bindings
for the old state and setting all bindings for the new state.  Previously
incremental adjustments were done at least some of the time.

This commit catches up with many commits on Subversion repository prior to
move to Git at a convenient point.