Remove a newline inserted before a PGN token when deleting token.
PGN text is displayed in tkinter.Text widgets configured to word wrap.
At regular intervals a newline is inserted before a fullmove for
performance reasons related to the volume of tagging.  A newline is
inserted before a non-move token in an attempt to make the layout look
nicer.  The PGN editor removes such newlines when the token after the
newline is deleted.