Ensure PGN token deletion does not make lines with too many moves.
Non-move tokens are usually on a line of their own.  When a token is
deleted keep one of the newlines, if necessary, to prevent creation of a
sequence of more than twenty fullmoves without a newline.

Other newlines are not adjusted, but after the game has been saved and
opened in a new widget the newline locations will probably have changed.