Fix editing of non-move token sequences including at edges of movetext.
Newlines are inserted and deleted, or not, when inserting or deleting
non-move movetext tokens so these appear on a line of their own without
extra blank lines between tokens.  The relevant tokens are '{}', '<>',
';\n', '\n%\n', and '(42. R)' an example of a Recursive Annotation
Variation immediately after insertion or before deletion.