Remove '\n%\n' from tokens visited by 'Next Comment' and similar.
These are PGN escaped lines.  They are still visited by 'Next Item'
action and similar.

One known problem exists at present.

If an escaped line is the first item in movetext the first move will
appear on the same line when given.  The move should be on a new line.
(This happens for the other non-move movetext items too.)