Avoid problems deleting RAVs for first move of game.
Problems, including program crashes, could occur if any RAV for the
first move of a game were deleted after a RAV for the first move of a
game was inserted, if the game had not been saved and re-opened between
insertion and deletion.

Dialogues are introduced to warn the situation is about to occur, and
offer the opportunity to not go there, and say what to do to get out of
the situation once there.