Allow display and edit of existing alternative RAV presentations.
RAVs are usually presented like 'd4 d5 c4 (g3) (Nf3)' where g3 and Nf3
are alternatives to c4.

The presentations 'd4 d5 c4 (Nf3 (g3))' and 'd4 d5 c4 ((g3) Nf3)', which
mean the same, are accepted but cannot be added.

These must not be confused with a layout like 'd4 d5 c4 (Nf3 Nf6 (g6))'
where g6 is an alternative to Nf6 and was accepted before this change.