Resolve pylint message missing-function-docstring.
These disagree with pydocstyle and, or, pycodestyle on the need for a
docstring in a _ClassName.method_name method.

Only 22 cases so give them all docstrings, even though the plan is convert
names to _ClassName.method_name or ClassName._method_name style where this
does not cause pydocstyle to give 'missing public docstring' messages: and
then remove the docstrings so they do not clutter generated documentation.

Bump chesstab version to 5.1.dev14 in setup.cfg.