Resolve some pylint invalid-name messages.
Bump chesstab version to 6.0.dev0 in setup.cfg.

The intention was to resolve invalid-name messages in 5.1.dev series but
it seems too much thought has to be given to whether a change belongs in
5.. or 6.. series.  There are at least several public function names for
which pylint gives this message by default, plus the redefined-builtin
message for the help module already deferred to 6.. series.