Fix problem introduced to gui.gameedit while fixing 'invalid-name's.
The way commit 9f69c443... resolved invalid-name messages looks unsound
and will be undone and repeated: not now because the close inspection
this module got revealed a sensible way of trying to break it up into
pieces which fit in boxes defined by the default plylint 'too-many-lines'
parameter.

Bump chesstab version to 6.0.dev14 in setup.cfg.