Resolve pylint too-many-lines message for gui.score module.
Module split into four: one for populating widget from PGN text, one for
event handlers and binding definitions, one for methods to implement
widget adjustments in response to events, and one for the AnalysisScore
class.

One method moved to the Game class, a subclass of Score, because it seems
to have been in the wrong place always.

Bump chesstab version to 6.0.dev16 in setup.cfg.