Resolve pylint import-error message for chessdptdu.
This adds a no-member message for chessdptdu, presumably undetectable
because of the import error.  The no-member message may have been present
in early reports, when attention was focused on the messages with several
hundred occurrences and before introduction of the 'shared' directory.

Change chesstab version to 6.0.dev73 in setup.cfg.