Undo the class-splitting done solely to reduce module size.
The stuff in modules scoreshow and scorewidget is put back in score.

The stuff in modules gameedit_widget, gameedit_handlers, gameedit_nonmove,
gameedit_misc, and gameedit_move, is put back in gamedit.

The stuff in modules chess_database, chess_import, and chess_widget, is
put back in chess.

Change chesstab version to 6.0.dev79 in setup.cfg.