Revert to text_count() workaround for tkinter.Text.count() method.
Attempts to delete any non-move item in PGN movetext, a numeric annotation
glyph for example, give an exception on the returned value from the
tkinter.Text.count() method.  Editing the item content is not affected.

The problem was introduced at ChessTab 5.0 after noticing the existence of
the tkinter method and using it since 4.0 in at least one place.