Add interface to Symas LMMD via solentware_base.
The major theme is adding wrappers for Symas LMMD read-only transactions
compatible with 'do-nothing' versions for the other database engines.

There is one unresolved known problem:

Scrolling a 'primary' database by dragging the scrollbar slider does not
work: the display is scrolled by one record rather than however many are
needed to fit the intended new position of the slider, and the slider is
positioned to fit the displayed records.  Scrolling this way in the other
database engines continues to work as expected.

Scrolling a 'secondary' database this way works as expected.

The 'primary' and 'secondary' terms loosely have the Berkeley DB meaning.