7.0.1 is new chesstab version.
Raise ScoreNoInitialPositionException, rather than TypeError, in
analysisscore.AnalysisScore.set_game_board to fit handling elsewhere.
The situation seems to have been expected but not implemented here.
Having seen the problem occur it has not proved possible to recreate
while testing change.

Fix two cases where RAVs are not inserted in exactly the right place:
a RAV for last move in game should not be placed on same line as Game
Termination Marker; a RAV for a black move should be placed before the
Move Number Indicator of the next white move, not after it.