Rework taking backups before deferred updates when importing PGN files.
Backups are not taken, by default, if the database engine supports
transactions: but appropriate subclasses can override and get backups
taken anyway, or not do transactions where allowed.

These backups are deleted if the deferred update succeeds.

Revert to backup with the appropriate decompression tool: this action
is not supported within chesstab.

Deferred updates takes more time with transactions than without.  But
backups take time too, and the time to do the deferred updates matters
as well.  Hopefully the expected cost (time) of the wrong choice is
minimized.

It became not possible to take backups in the ChessTab program for the
DPT database engine, but became possible to do deferred updates with the
tkinter/tcl interface to Berkeley DB without the workaround in samples.

Some hints by pylint, pycodestyle and pydocstyle are implemented.

EditPGNToplevel and Alldu classes are added to ignored-classes in
pylint.conf.

The archivedu module is moved to solentware-base project because it no
longer has any chesstab specific code.

7.1.dev2 is new chesstab version.