Improve data and index space allocation for DPT database engine.
The space allocation is always a guess without the possibility of instant
correction if wrong: the improvent is the guess is less likely to be too
small and not far too large.  See solentware-base for details.

Database interface naming convention for modules and classes is changed
to that used in version 2.0.0.dev0 of chesscalc.

The changes in the shared.alldu module are probably significant for future
improvents.

New dependency version 5.2.dev0 of solentware-base.

New dependency version 2.1.3.dev0 of chessql.

New dependency version 2.4.dev0 of pgn-read.

New version 7.1.4.dev0 of chesstab.