All PGN imports are done within transactions.
This is slower than before, especially with the DPT database engine, but
imports can be interrupted at any point and a repeat of the import will
figure what has been done already and restart immediately after the most
recent commit.  This goes a long way to meeting issue 2 but it is
impossible to provide user access to the database while running any part
of import.

The source PGN file name, and game number within file, are stored on the
database so restart from most recent commit can be done.

Thus this version of ChessTab cannot access databases created by earlier
versions, and vice-versa.  Export and import between them is possible.

Issue 1 is resolved.
Issue 3 is resolved.