Resolve issues 4 and 5.
A "*" game termination marker is added if no game termination marker is
present, and the preceding movetext is treated as invalid even if it is
valid and is wrapped in a "{}" comment.

Games and repertoires with "" PGN tag values which would become index
keys are treated as invalid even if they are otherwise valid.  Games
are put on the error index and repertoire update is not allowed in
these cases.