Adjust initial import log messages and add temporary file copy of log.
The initial messages are intended to guide choice of import action:
direct to database or by sorting to files and merge with an index dump
from database.

A temporary file copy of log is made because large imports take hours or
days.  An option to make a permanent copy on completion already exists.