Remove index references for deleted games and avoid zero length keys.
The zero length keys were for games not imported from a PGN file and
edited games.  '/' is now used as file name key for these games to
allow storage on Symas LMDB via lmdb from PyPI.

Remove index references if a game displayed by 'F11' action on a list
of games is deleted via the popup menu for the game.

The 'Delete' action on a list of games already removed the index
references.

New version 8.1.dev2 of chesstab.