Evaluate CQL queries with the CQL program.
CQL queries are evaluated by running the CQL program and updating
indicies from the CQL program output.

All code which evaluates CQL queries within ChessTab is removed.  The
dependency on ChessQL to parse CQL queries is removed.

The index of piece location at each move is discarded.  An index of
squares visited by each piece during the game is added (but the
intention is to remove it).

Games are indexed by all the PGN tags mentioned in the PGN specification
plus some others which seem in common use.

When a database is opened after an import was interrupted a prompt to
resume the import is done.

When a database is opened after a CQL query evaluation was interrupted
the evaluation is resumed.

Module chesstab.lmdb.database is modified to cope with Github issue
jnwatson/py-lmdb 388.

Several version changes were convenient while implementing these
features in a branch in the local Subversion repository.  None of the
intermediate points are committed in the Git repository (so dev3
through dev7 are missing).

New version 8.1.dev8 of chesstab.