Reverse changes for faster creation of PGN-like files for CQL scans.
These were in commit '4500d067..'.

The changes broke import of games where three or more pieces of a type
could move to a square and full disambiguation, like Qb1c2, was needed.
The whole import would succeed but affected games were marked error in
direct imports or would cause a program crash on attempting to display
the game when imported by merge.

The problem did not occur where full disambiguation was used but not
needed.  Games in the LiChess January 2017 download exposed this.

Adjust exporters of PGN files for CQL evaluation to fit.