Fit chesstab to renamed objects in pgn-read.
The slower version of import games from PGN files is back.

It is hoped the changes in pgn-read will make it easier to speed up
game import.

A PGN parser for a modified game record structure is added: it may get
used to speed up game import if the pgn-read changes do not do enough.

New version 2.5.dev2 of dependency pgn-read.

New version 8.1.dev10 of chesstab.